Gomigrator V1 (#1)
Initial build of a tool to help manage migrations in go projects.
This includes some components that will help run a defined set of
migrations for any go application. A set of migrations and a db
connection can be passed to the migrator which will then handle
the logic of applying any unapplied migrations and rolling them
back.

It currently supports postgres only.

It's extensively unit tested with datadogs sqlmock library.
Integration tests will be coming soon.

It includes github actions to run tests in CI for pull requests.
The tests will be run on golang 1.15, 1.16 and 1.17.

The Makefile and docker-compose files are used to run the tests
locally on go 1.17.