Add integration tests (#2)
Added integration tests for postgres. Locally the docker compose
environment spins up a postgres container and supplies env vars to
the container that runs integration tests. This is all handled by
the makefile.

Moved the docker-compose file to avoid confusion with the .env file
as it is only used for the docker-compose file and not the code.

Ran go fmt on most of the code.

Added golang ci lint, for local and ci.